fix(llm):portmapping of model network only works when network mode is hostlocal
